Refactor profile editor structure
Refs #159
